Add removeUser() and getName()
